Move file.helper and interfaces to the model folder
