+++ visualization/app/codeCharta/state/metric.service.ts
@@ -1,4 +1,4 @@
-import { BlacklistItem, BlacklistType, CodeMapNode, FileState, MetricData, AttributeTypeValue, AttributeTypes } from "../codeCharta.model"
+import { BlacklistItem, BlacklistType, CodeMapNode, MetricData, AttributeTypeValue, AttributeTypes } from "../codeCharta.model"
 import { hierarchy, HierarchyNode } from "d3"
 import { IRootScopeService } from "angular"
 import { CodeMapHelper } from "../util/codeMapHelper"
@@ -6,7 +6,8 @@ import { BlacklistService, BlacklistSubscriber } from "./store/fileSettings/blac
 import { StoreService } from "./store.service"
 import { FilesService, FilesSelectionSubscriber } from "./store/files/files.service"
 import { AttributeTypesSubscriber, AttributeTypesService } from "./store/fileSettings/attributeTypes/attributeTypes.service"
-import { fileStatesAvailable, getVisibleFileStates } from "./store/files/files.helper"
+import { fileStatesAvailable, getVisibleFileStates } from "../model/files/files.helper"
+import { FileState } from "../model/files/files"
 
 export interface MetricServiceSubscriber {
 	onMetricDataAdded(metricData: MetricData[])
