+++ visualization/app/codeCharta/ui/codeMap/threeViewer/threeOrbitControlsService.ts
@@ -9,7 +9,7 @@ import {
 	UnfocusNodeSubscriber
 } from "../../../state/store/dynamicSettings/focusedNodePath/focusedNodePath.service"
 import { FilesService, FilesSelectionSubscriber } from "../../../state/store/files/files.service"
-import { FileState } from "../../../codeCharta.model"
+import { FileState } from "../../../model/files/files"
 
 export interface CameraChangeSubscriber {
 	onCameraChanged(camera: PerspectiveCamera)
