+++ visualization/app/codeCharta/ui/colorSettingsPanel/colorSettingsPanel.component.ts
@@ -18,8 +18,8 @@ import {
 	InvertColorRangeSubscriber
 } from "../../state/store/appSettings/invertColorRange/invertColorRange.service"
 import { FilesService, FilesSelectionSubscriber } from "../../state/store/files/files.service"
-import { FileState } from "../../codeCharta.model"
-import { isDeltaState } from "../../state/store/files/files.helper"
+import { isDeltaState } from "../../model/files/files.helper"
+import { FileState } from "../../model/files/files"
 
 export class ColorSettingsPanelController
 	implements FilesSelectionSubscriber, InvertDeltaColorsSubscriber, WhiteColorBuildingsSubscriber, InvertColorRangeSubscriber {
